:bug: fix: cleaner selected bug solved
